+++ covalent_ui/webapp/src/components/graph/ElectronNode.js
@@ -20,15 +20,10 @@
  * Relief from the License may be granted by purchasing a commercial license.
  */
 
-import {
-  Grid,
-  Paper,
-  Tooltip,
-  tooltipClasses,
-  Typography,
-} from '@mui/material'
+import { Grid, Paper, Tooltip, tooltipClasses, Typography } from '@mui/material'
 import { styled } from '@mui/material/styles'
 import { Handle } from 'react-flow-renderer'
+import CopyButton from '../common/CopyButton'
 import { truncateMiddle, nodeLabelIcon, statusIcon } from '../../utils/misc'
 
 export const NODE_TEXT_COLOR = 'rgba(250, 250, 250, 0.6)'
@@ -101,11 +96,13 @@ const ElectronNode = ({
           title={
             data.hideLabels ? (
               <>
-                <Typography color="inherit">name : {data.fullName}</Typography>
-                <Typography color="inherit">
+                <Typography sx={{ fontSize: '0.75rem' }} color="inherit">
+                  name : {data.fullName}
+                </Typography>
+                <Typography sx={{ fontSize: '0.75rem' }} color="inherit">
                   executor : {data.executor}
                 </Typography>
-                <Typography color="inherit">
+                <Typography sx={{ fontSize: '0.75rem' }} color="inherit">
                   node_id : {data.node_id}
                 </Typography>
               </>
@@ -149,13 +146,29 @@ const ElectronNode = ({
             />
             {nodeLabelIcon(data.nodeType)}
             {statusIcon(data.status)}
-            <Typography sx={{ fontSize: 14, mb: 0.3 }}>{data.label}</Typography>
+            <Typography sx={{ fontSize: 14, mb: 0.3, mt: 0.3 }}>
+              {data.label}
+            </Typography>
             <Handle
               data-testid="sourcehandleelectronNode"
               type="source"
               position={sourcePosition}
               isConnectable={isConnectable}
             />
+            {data.nodeType === 'sublattice' && (
+              <CopyButton
+                sx={{ ml: 1, color: 'text.tertiary', pt: 0.3 }}
+                fontSize="10"
+                content={data.sublattices_id}
+                size="small"
+                className="copy-btn"
+                title="Copy ID"
+                width={17}
+                height={17}
+                isBorderPresent={true}
+                borderRadius="4px"
+              />
+            )}
           </Paper>
         </ElectronTooltip>
         {!data.hideLabels ? (