+++ covalent_ui/webapp/src/components/settings/SettingsCard.js
@@ -53,7 +53,7 @@ import FormControl from '@mui/material/FormControl'
 import FormLabel from '@mui/material/FormLabel'
 import { settingsResults, updateSettings } from '../../redux/settingsSlice'
 import { useDispatch, useSelector } from 'react-redux'
-import _, { capitalize } from 'lodash'
+import _ from 'lodash'
 import Skeleton from '@mui/material/Skeleton'
 import { ReactComponent as closeIcon } from '../../assets/close.svg'
 import { toggleLatticeDrawer } from '../../redux/popupSlice'
@@ -145,7 +145,7 @@ const SettingsCard = () => {
       } else if (action.type === updateSettings.rejected.type) {
         setOpenSnackbar(true)
         setSnackbarMessage(
-          'Something went wrong and could not settings updated!'
+          'Something went wrong and settings could not be updated.'
         )
       }
     })
@@ -219,7 +219,7 @@ const SettingsCard = () => {
         } else if (action.type === updateSettings.rejected.type) {
           setOpenSnackbar(true)
           setSnackbarMessage(
-            'Something went wrong and could not settings updated!'
+            'Something went wrong and settings could not be updated.'
           )
         }
       })
@@ -405,6 +405,9 @@ const SettingsCard = () => {
                     border: '1px solid #303067',
                     borderRadius: '60px',
                     mb: 3,
+                    '@media (max-width: 1290px)': {
+                      width: '258px',
+                    },
                   }}
                   disableUnderline
                   value={searchKey}
@@ -437,26 +440,15 @@ const SettingsCard = () => {
                   onChange={(e) => handleInputChange(e)}
                   placeholder="Search"
                 />
-                <Typography
-                  variant="h6"
-                  component="h6"
-                  sx={(theme) => ({
-                    color: theme.palette.primary.white,
-                    fontWeight: 'bold',
-                    mb: 2,
-                  })}
-                >
-                  {capitalize(accName)}
-                </Typography>
                 {_.map(clientDetail, function (menuValue, menuKey) {
                   return (
                     <StyledList sx={{ pb: 0, pt: 0 }} key={menuKey}>
-                      <ListItem disablePadding>
+                      <ListItem disablePadding sx={{ lineHeight: '18px' }}>
                         <ListItemButton
                           onClick={
                             isChildHasList
                               ? () => handleClick(menuValue)
-                              : () => {}
+                              : () => { }
                           }
                           sx={{
                             right: '0px',
@@ -505,7 +497,7 @@ const SettingsCard = () => {
                       {_.map(subMenu, function (value, key) {
                         return (
                           <StyledList sx={{ pb: 0, pt: 0 }} key={key}>
-                            <ListItem disablePadding>
+                            <ListItem disablePadding sx={{ lineHeight: '18px' }}>
                               <ListItemButton
                                 sx={{ pl: 7, pt: 0.3, pb: 0.3 }}
                                 onClick={() => handleSubmenuClick(subMenu, key)}
@@ -529,27 +521,15 @@ const SettingsCard = () => {
               </Box>
 
               <Box>
-                <Typography
-                  variant="h6"
-                  component="h6"
-                  sx={(theme) => ({
-                    color: theme.palette.primary.white,
-                    fontWeight: 'bold',
-                    mb: 2,
-                    mt: 2,
-                  })}
-                >
-                  {capitalize(serverName)}
-                </Typography>
                 {_.map(serverDetail, function (menuValue, menuKey) {
                   return (
                     <StyledList sx={{ pb: 0, pt: 0 }} key={menuKey}>
-                      <ListItem disablePadding>
+                      <ListItem disablePadding sx={{ lineHeight: '18px' }}>
                         <ListItemButton
                           onClick={
                             isChildHasList
                               ? () => handleClick(menuValue)
-                              : () => {}
+                              : () => { }
                           }
                           sx={{
                             right: '0px',
@@ -600,6 +580,7 @@ const SettingsCard = () => {
                   fontWeight: 'bold',
                   fontSize: '24px',
                   color: (theme) => theme.palette.primary.white,
+                  lineHeight: '31px'
                 }}
               >
                 {formatUnderscoreConcatenatedString(resultKey)}
@@ -616,7 +597,7 @@ const SettingsCard = () => {
                           height: 550,
                         },
                         '@media (min-height: 1000px)': {
-                          height: 750,
+                          height: 730,
                         },
                       }}
                     >
@@ -624,13 +605,15 @@ const SettingsCard = () => {
                         return (
                           <Box sx={{ mb: 3 }} key={key}>
                             {_.isObject(value) ? (
-                              <Box key={key} id={key}>
+                              <Box key={key} id={key} sx={{ mb: '50px' }}>
                                 <Typography
                                   variant="h6"
                                   component="h6"
                                   sx={(theme) => ({
                                     color: theme.palette.primary.light,
                                     fontWeight: 'bold',
+                                    lineHeight: '16px',
+                                    fontSize:'16px'
                                   })}
                                 >
                                   {formatUnderscoreConcatenatedString(key)}